Halt on 500 level errors
If a 500 level error is seen, halt omicron-stress if
`--server-errors-fatal` is provided.

If an antagonist sees an error, it should emit a `warn` message: later,
when collecting the errors, if an error is considered fatal it should be
emit in an `error` message. This will allow easier skimming of the log
output.